test: share FastAPI stubs
