+++ tests/test_allowed_origins.py
@@ -1,3 +1,5 @@
+import sys, os
+sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 import os
 import sys
 import json
@@ -11,61 +13,16 @@ psutil_stub.disk_usage = lambda path: types.SimpleNamespace(percent=0.0)
 psutil_stub.getloadavg = lambda: (0.0, 0.0, 0.0)
 sys.modules.setdefault("psutil", psutil_stub)
 
-fastapi_stub = types.ModuleType("fastapi")
+from tests.test_helpers import (
+    fastapi_stub,
+    cors_stub,
+    resp_stub,
+    pydantic_stub,
+    install_stubs,
+    FakeApp,
+)
 
-class FakeApp:
-    def __init__(self):
-        self.added = []
-
-    def add_middleware(self, *a, **kw):
-        self.added.append((a, kw))
-
-    def on_event(self, *a, **kw):
-        return lambda f: f
-
-    def post(self, *a, **kw):
-        return lambda f: f
-
-    def get(self, *a, **kw):
-        return lambda f: f
-
-    def delete(self, *a, **kw):
-        return lambda f: f
-
-    def websocket(self, *a, **kw):
-        return lambda f: f
-
-fastapi_stub.FastAPI = lambda: FakeApp()
-fastapi_stub.UploadFile = object
-fastapi_stub.File = lambda *a, **kw: None
-fastapi_stub.WebSocket = object
-fastapi_stub.WebSocketDisconnect = type("WebSocketDisconnect", (Exception,), {})
-
-class HTTPException(Exception):
-    pass
-fastapi_stub.HTTPException = HTTPException
-
-pydantic_stub = types.ModuleType("pydantic")
-class BaseModel:
-    pass
-pydantic_stub.BaseModel = BaseModel
-
-cors_stub = types.ModuleType("fastapi.middleware.cors")
-cors_stub.CORSMiddleware = type("CORSMiddleware", (), {})
-
-resp_stub = types.ModuleType("fastapi.responses")
-class DummyHTMLResponse:
-    def __init__(self, text, status_code=200):
-        self.text = text
-        self.status_code = status_code
-        self.called = False
-
-    async def __call__(self, scope, receive, send):
-        self.called = True
-        await send({"status": self.status_code, "body": self.text})
-
-resp_stub.HTMLResponse = DummyHTMLResponse
-resp_stub.FileResponse = object
+install_stubs()
 
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
