+++ tests/test_auth_middleware.py
@@ -1,78 +1,21 @@
+import sys, os
+sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 import asyncio
 import json
 import sys
 import os
-import types
 import pytest
 
-# Stub FastAPI and related pieces similar to other tests
-fastapi_stub = types.ModuleType("fastapi")
+from tests.test_helpers import (
+    fastapi_stub,
+    cors_stub,
+    resp_stub,
+    pydantic_stub,
+    install_stubs,
+    FakeApp,
+)
 
-class FakeApp:
-    def __init__(self):
-        self.called = False
-
-    async def __call__(self, scope, receive, send):
-        self.called = True
-        await send({"done": True})
-
-    def add_middleware(self, *a, **kw):
-        pass
-
-    def on_event(self, *a, **kw):
-        return lambda f: f
-
-    def post(self, *a, **kw):
-        return lambda f: f
-
-    def get(self, *a, **kw):
-        return lambda f: f
-
-    def delete(self, *a, **kw):
-        return lambda f: f
-
-    def websocket(self, *a, **kw):
-        return lambda f: f
-
-fastapi_stub.FastAPI = lambda: FakeApp()
-fastapi_stub.UploadFile = object
-fastapi_stub.File = lambda *a, **kw: None
-fastapi_stub.WebSocket = object
-fastapi_stub.WebSocketDisconnect = type("WebSocketDisconnect", (Exception,), {})
-
-class HTTPException(Exception):
-    pass
-
-fastapi_stub.HTTPException = HTTPException
-sys.modules.setdefault("fastapi", fastapi_stub)
-
-cors_stub = types.ModuleType("fastapi.middleware.cors")
-cors_stub.CORSMiddleware = object
-sys.modules.setdefault("fastapi.middleware.cors", cors_stub)
-
-resp_stub = types.ModuleType("fastapi.responses")
-
-class DummyHTMLResponse:
-    def __init__(self, text, status_code=200):
-        self.text = text
-        self.status_code = status_code
-        self.called = False
-
-    async def __call__(self, scope, receive, send):
-        self.called = True
-        await send({"status": self.status_code, "body": self.text})
-
-resp_stub.HTMLResponse = DummyHTMLResponse
-resp_stub.FileResponse = object
-sys.modules.setdefault("fastapi.responses", resp_stub)
-
-pydantic_stub = types.ModuleType("pydantic")
-
-class BaseModel:
-    pass
-
-pydantic_stub.BaseModel = BaseModel
-sys.modules.setdefault("pydantic", pydantic_stub)
+install_stubs()
