+++ tests/test_wordlist_db.py
@@ -1,68 +1,23 @@
+import sys, os
+sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 import asyncio
 import sys
 import os
-import types
 import sqlite3
 import json
 import base64
 import gzip
 
-# Stub modules similar to other server tests
-fastapi_stub = types.ModuleType("fastapi")
+from tests.test_helpers import (
+    fastapi_stub,
+    cors_stub,
+    resp_stub,
+    pydantic_stub,
+    install_stubs,
+    FakeApp,
+)
 
-
-class FakeApp:
-    def add_middleware(self, *a, **kw):
-        pass
-
-    def on_event(self, *a, **kw):
-        return lambda f: f
-
-    def post(self, *a, **kw):
-        return lambda f: f
-
-    def get(self, *a, **kw):
-        return lambda f: f
-
-    def delete(self, *a, **kw):
-        return lambda f: f
-
-    def websocket(self, *a, **kw):
-        return lambda f: f
-
-
-fastapi_stub.FastAPI = lambda: FakeApp()
-fastapi_stub.UploadFile = object
-fastapi_stub.File = lambda *a, **kw: None
-fastapi_stub.WebSocket = object
-fastapi_stub.WebSocketDisconnect = type("WebSocketDisconnect", (), {})
-
-
-class HTTPException(Exception):
-    pass
-
-
-fastapi_stub.HTTPException = HTTPException
-sys.modules.setdefault("fastapi", fastapi_stub)
-
-cors_stub = types.ModuleType("fastapi.middleware.cors")
-cors_stub.CORSMiddleware = object
-sys.modules.setdefault("fastapi.middleware.cors", cors_stub)
-
-resp_stub = types.ModuleType("fastapi.responses")
-resp_stub.HTMLResponse = object
-resp_stub.FileResponse = object
-sys.modules.setdefault("fastapi.responses", resp_stub)
-
-pydantic_stub = types.ModuleType("pydantic")
-
-
-class BaseModel:
-    pass
-
-
-pydantic_stub.BaseModel = BaseModel
-sys.modules.setdefault("pydantic", pydantic_stub)
+install_stubs()
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 sys.path.insert(0, os.path.join(os.path.dirname(os.path.dirname(__file__)), "Server"))
